+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - **[Breaking]** Try and As conversion helpers are removed in favor of using
   other cast libraries.
 - **[Breaking]** Removed `Value.IsDefault` method.
+- **[Breaking]** Removed Load* functions.
+- **[Breaking]** Removed NewYAMLProviderFromReader* functions.
 
 ## v1.0.0-rc1 (26 Jun 2017)